pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-model-rag"
version = "0.1.0"
description = "Production-oriented multi-model RAG project scaffold"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "google-genai>=1.0.0",
  "openai>=1.0.0",
  "psycopg[binary]>=3.2.0",
  "sqlalchemy>=2.0.30",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "mypy>=1.10.0",
  "pytest>=8.2.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true

[tool.setuptools.packages.find]
include = ["apps*", "packages*"]
exclude = ["tests*", "docs*", "infra*", "datasets*", "migrations*"]
